requirements.txt
APScheduler==3.10.4
asyncio==3.4.3
dnspython==2.4.2
discord.py>=1.7.3
emojis==0.7.0
flask==3.0.0
matplotlib==3.8.2
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
tzlocal==5.2
waitress==2.1.2